+++ src/reputation/index.ts
@@ -1,11 +1,17 @@
 export { appendResponse } from './appendResponse.js'
 export { getClients } from './getClients.js'
+export { getIdentityRegistry } from './getIdentityRegistry.js'
 export { getLastIndex } from './getLastIndex.js'
 export { getResponseCount } from './getResponseCount.js'
 export { getSummary } from './getSummary.js'
+export { getVersion } from './getVersion.js'
 export { giveFeedback } from './giveFeedback.js'
+export type { FeedbackRevokedResult } from './parseFeedbackRevokedReceipt.js'
+export { parseFeedbackRevokedReceipt } from './parseFeedbackRevokedReceipt.js'
 export type { GiveFeedbackResult } from './parseGiveFeedbackReceipt.js'
 export { parseGiveFeedbackReceipt } from './parseGiveFeedbackReceipt.js'
+export type { ResponseAppendedResult } from './parseResponseAppendedReceipt.js'
+export { parseResponseAppendedReceipt } from './parseResponseAppendedReceipt.js'
 export { readAllFeedback } from './readAllFeedback.js'
 export { readAllFeedbackBatched } from './readAllFeedbackBatched.js'
 export { readFeedback } from './readFeedback.js'
@@ -16,9 +22,11 @@ export type {
   Feedback,
   FeedbackEntry,
   GetClientsParameters,
+  GetIdentityRegistryParameters,
   GetLastIndexParameters,
   GetResponseCountParameters,
   GetSummaryParameters,
+  GetVersionParameters,
   GiveFeedbackParameters,
   ReadAllFeedbackBatchedParameters,
   ReadAllFeedbackParameters,